[refactor] 유사한 기능을 하는 Controller 및 Service 통합
- Board 관련 API BoardController로 통합
- Board CRUD 관련 기능 BoardService로 통합
- Board 리스트 조회 관련 기능 BoardListService로 통합
- Memo 관련 기능 MemoController 및 MemoService로 통합
- Theme 관련 기능 ThemeController 및 ThemeService로 통합
