+++ src/main/java/com/boardwe/boardwe/controller/ThemeController.java
@@ -2,7 +2,7 @@ package com.boardwe.boardwe.controller;
 
 import com.boardwe.boardwe.dto.res.BoardThemeSelectResponseDto;
 import com.boardwe.boardwe.dto.res.MemoThemeSelectResponseDto;
-import com.boardwe.boardwe.service.ThemeSelectService;
+import com.boardwe.boardwe.service.ThemeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,17 +13,17 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-public class ThemeSelectController {
+public class ThemeController {
 
-    private final ThemeSelectService themeSelectService;
+    private final ThemeService themeService;
 
     @GetMapping("/board-themes")
     public ResponseEntity<List<BoardThemeSelectResponseDto>> getBoardThemes() {
-        return ResponseEntity.ok(themeSelectService.getPublicThemes());
+        return ResponseEntity.ok(themeService.getPublicThemes());
     }
 
     @GetMapping("/board/{boardCode}/memo-themes")
     public ResponseEntity<List<MemoThemeSelectResponseDto>> getMemoThemes(@PathVariable String boardCode) {
-        return ResponseEntity.ok(themeSelectService.getMemoThemesOfBoard(boardCode));
+        return ResponseEntity.ok(themeService.getMemoThemesOfBoard(boardCode));
     }
 }
