--- src/main/java/com/boardwe/boardwe/service/BoardCreateService.java
@@ -1,9 +1,0 @@
-package com.boardwe.boardwe.service;
-
-import com.boardwe.boardwe.dto.req.BoardCreateRequestDto;
-import com.boardwe.boardwe.dto.res.BoardCreateResponseDto;
-
-public interface BoardCreateService {
-
-    BoardCreateResponseDto createBoard(BoardCreateRequestDto requestDto);
-}
