--- src/main/java/com/boardwe/boardwe/service/BoardDeleteService.java
@@ -1,7 +1,0 @@
-package com.boardwe.boardwe.service;
-
-import com.boardwe.boardwe.dto.req.BoardDeleteRequestDto;
-
-public interface BoardDeleteService {
-    void deleteBoard(BoardDeleteRequestDto requestDto, String boardCode);
-}
