+++ src/main/java/com/boardwe/boardwe/service/BoardListService.java
@@ -1,0 +1,10 @@
+package com.boardwe.boardwe.service;
+
+import java.util.List;
+
+import com.boardwe.boardwe.dto.res.BoardSearchResultResponseDto;
+
+public interface BoardListService {
+    List<BoardSearchResultResponseDto> selectHotBoards();
+    List<BoardSearchResultResponseDto> selectRecommendBoards();
+}
