--- src/main/java/com/boardwe/boardwe/service/BoardMemoSearchService.java
@@ -1,7 +1,0 @@
-package com.boardwe.boardwe.service;
-
-import com.boardwe.boardwe.dto.res.BoardMemoSearchResponseDto;
-
-public interface BoardMemoSearchService {
-    BoardMemoSearchResponseDto searchMemo(String boardCode, String query);
-}
