--- src/main/java/com/boardwe/boardwe/service/BoardSelectService.java
@@ -1,7 +1,0 @@
-package com.boardwe.boardwe.service;
-
-import com.boardwe.boardwe.dto.res.BoardSelectResponseDto;
-
-public interface BoardSelectService {
-    BoardSelectResponseDto getBoard(String boardCode);
-}
