--- src/main/java/com/boardwe/boardwe/service/MemoDeleteService.java
@@ -1,7 +1,0 @@
-package com.boardwe.boardwe.service;
-
-import com.boardwe.boardwe.dto.req.MemoDeleteRequestDto;
-
-public interface MemoDeleteService {
-    void deleteMemo(MemoDeleteRequestDto memoDeleteRequestDto, String boardCode);
-}
