+++ src/main/java/com/boardwe/boardwe/service/MemoService.java
@@ -1,0 +1,12 @@
+package com.boardwe.boardwe.service;
+
+import com.boardwe.boardwe.dto.req.MemoCreateRequestDto;
+import com.boardwe.boardwe.dto.req.MemoDeleteRequestDto;
+import com.boardwe.boardwe.dto.res.MemoSearchResponseDto;
+import com.boardwe.boardwe.dto.res.MemoCreateResponseDto;
+
+public interface MemoService {
+    MemoCreateResponseDto createMemo(MemoCreateRequestDto memoCreateRequestDto, String boardCode);
+    void deleteMemo(MemoDeleteRequestDto memoDeleteRequestDto, String boardCode);
+    MemoSearchResponseDto searchMemo(String boardCode, String query);
+}
