--- src/main/java/com/boardwe/boardwe/service/RecommendBoardListSelectService.java
@@ -1,9 +1,0 @@
-package com.boardwe.boardwe.service;
-
-import java.util.List;
-
-import com.boardwe.boardwe.dto.res.BoardSearchResultResponseDto;
-
-public interface RecommendBoardListSelectService {
-    List<BoardSearchResultResponseDto> getBoardList();
-}
