--- src/main/java/com/boardwe/boardwe/service/WelcomeBoardService.java
@@ -1,7 +1,0 @@
-package com.boardwe.boardwe.service;
-
-import com.boardwe.boardwe.dto.res.WelcomeBoardResponseDto;
-
-public interface WelcomeBoardService {
-    WelcomeBoardResponseDto getWelcomBoardDto(String boardCode);
-}
