+++ src/main/java/com/boardwe/boardwe/service/impl/MemoServiceImpl.java
@@ -1,15 +1,21 @@
 package com.boardwe.boardwe.service.impl;
 
-import com.boardwe.boardwe.dto.res.inner.MemoSelectResponseDto;
-import com.boardwe.boardwe.dto.res.BoardMemoSearchResponseDto;
+import com.boardwe.boardwe.dto.req.MemoCreateRequestDto;
+import com.boardwe.boardwe.dto.req.MemoDeleteRequestDto;
 import com.boardwe.boardwe.dto.res.BoardThemeSelectResponseDto;
+import com.boardwe.boardwe.dto.res.MemoCreateResponseDto;
+import com.boardwe.boardwe.dto.res.MemoSearchResponseDto;
 import com.boardwe.boardwe.dto.res.MemoThemeSelectResponseDto;
+import com.boardwe.boardwe.dto.res.inner.MemoSelectResponseDto;
 import com.boardwe.boardwe.entity.*;
 import com.boardwe.boardwe.exception.custom.BoardNotFoundException;
+import com.boardwe.boardwe.exception.custom.MemoNotFoundException;
+import com.boardwe.boardwe.exception.custom.MemoThemeNotFoundException;
+import com.boardwe.boardwe.exception.custom.MemoWithInvalidBoardException;
 import com.boardwe.boardwe.repository.BoardRepository;
 import com.boardwe.boardwe.repository.MemoRepository;
 import com.boardwe.boardwe.repository.MemoThemeRepository;
-import com.boardwe.boardwe.service.BoardMemoSearchService;
+import com.boardwe.boardwe.service.MemoService;
 import com.boardwe.boardwe.type.BackgroundType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,18 +23,57 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
-public class BoardMemoSearchServiceImpl implements BoardMemoSearchService {
+@Transactional(readOnly = true)
+public class MemoServiceImpl implements MemoService {
 
-    private final BoardRepository boardRepository;
     private final MemoRepository memoRepository;
+    private final BoardRepository boardRepository;
     private final MemoThemeRepository memoThemeRepository;
 
     @Override
-    public BoardMemoSearchResponseDto searchMemo(String boardCode, String query) {
+    @Transactional
+    public MemoCreateResponseDto createMemo(MemoCreateRequestDto memoCreateRequestDto, String boardCode) {
+        Board board = boardRepository.findByCode(boardCode).orElseThrow(BoardNotFoundException::new);
+
+        MemoTheme memoTheme = memoThemeRepository.findById(memoCreateRequestDto.getMemoThemeId()).orElseThrow(MemoThemeNotFoundException::new);
+
+        Memo memo = Memo.builder()
+                .board(board)
+                .memoTheme(memoTheme)
+                .content(memoCreateRequestDto.getMemoContent())
+                .build();
+        memoRepository.save(memo);
+
+        return MemoCreateResponseDto.builder().openStartTime(board.getOpenStartTime()).build();
+    }
+
+    @Override
+    @Transactional
+    public void deleteMemo(MemoDeleteRequestDto memoDeleteRequestDto, String boardCode) {
+        Board board = boardRepository.findByCode(boardCode).orElseThrow(BoardNotFoundException::new);
+        /**
+         * TODO
+         * Memo & Board Fetch Join 필요
+         * Session ID 적용 필요
+         */
+        for (Long memoId : memoDeleteRequestDto.getMemoIds()) {
+            Memo memo = memoRepository.findById(memoId).orElseThrow(MemoNotFoundException::new);
+
+            if(Objects.equals(memo.getBoard().getCode(), boardCode)){
+                memoRepository.deleteById(memoId);
+            }else{
+                throw new MemoWithInvalidBoardException();
+            }
+
+        }
+    }
+
+    @Override
+    public MemoSearchResponseDto searchMemo(String boardCode, String query) {
         Board board = boardRepository.findByCode(boardCode).orElseThrow(BoardNotFoundException::new);
 
         List<Memo> memos = memoRepository.findByBoardAndContentContains(board, query);
@@ -55,12 +100,12 @@ public class BoardMemoSearchServiceImpl implements BoardMemoSearchService {
             memoDtos.add(memoDto);
         }
 
-        return BoardMemoSearchResponseDto.builder()
+        return MemoSearchResponseDto.builder()
                 .theme(boardThemeDto)
                 .memos(memoDtos)
                 .build();
-
     }
+
     private List<MemoThemeSelectResponseDto> getMemoThemeResponseDtos(BoardTheme boardTheme) {
         List<MemoTheme> memoThemes = memoThemeRepository.findByBoardThemeId(boardTheme.getId());
         List<MemoThemeSelectResponseDto> memoThemeSelectResponseDtos = new ArrayList<>();
